moved test_structure migration and constructor tests to separate files for readability
