+++ particle_structs/test/test_migrate.cpp
@@ -1,0 +1,203 @@
+#include <particle_structs.hpp>
+#include "read_particles.hpp"
+
+int testMigration(const char* name, PS* structure) {
+  printf("testMigration %s, rank %d\n", name, comm_rank);
+
+  int fails = 0;
+  kkLidView failures("fails", 1);
+
+  kkLidView new_element("new_element", structure->capacity());
+  kkLidView new_process("new_process", structure->capacity());
+
+  int num_ptcls = structure->nPtcls();
+  //Send even particles one process to the right
+  auto pids = structure->get<0>();
+  auto rnks = structure->get<3>();
+  int local_rank = comm_rank;
+  int local_csize = comm_size;
+  int num_elems = structure->nElems();
+  auto sendRight = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
+    if (mask) {
+      new_element(p) = e;
+      if (e == num_elems - 1) {
+        new_process(p) = (local_rank + 1) % local_csize;
+      }
+      else
+        new_process(p) = local_rank;
+      rnks(p) = local_rank;
+    }
+    else {
+      new_element(p) = e;
+      new_process(p) = local_rank;
+    }
+  };
+  ps::parallel_for(structure, sendRight, "sendRight");
+  structure->migrate(new_element, new_process);
+
+  pids = structure->get<0>();
+  rnks = structure->get<3>();
+  auto checkPostMigrate = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
+    if (mask) {
+      const int pid = pids(p);
+      const int rank = rnks(p);
+      if (e == num_elems - 1 && rank == local_rank) {
+        printf("[ERROR] Failed to send particle %d on rank %d\n",
+               pid, local_rank);
+        failures(0) = 1;
+      }
+      if (e != 0 && rank != local_rank) {
+        printf("[ERROR] Incorrectly received particle %d from rank %d to rank %d in element %d\n", pid, rank, local_rank, e);
+        failures(0) = 1;
+      }
+    }
+  };
+  if (comm_size > 1)
+    ps::parallel_for(structure, checkPostMigrate, "checkPostMigrate");
+  fails += ps::getLastValue<lid_t>(failures);
+
+  //Make a distributor
+  int neighbors[3];
+  neighbors[0] = comm_rank;
+  neighbors[1] = (comm_rank - 1 + comm_size) % comm_size;
+  neighbors[2] = (comm_rank + 1) % comm_size;
+  ps::Distributor<typename PS::memory_space> dist(std::min(comm_size, 3), neighbors);
+
+  new_element = kkLidView("new_element", structure->capacity());
+  new_process = kkLidView("new_process", structure->capacity());
+  auto sendBack = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
+    new_element(p) = e;
+    new_process(p) = rnks(p);
+  };
+  ps::parallel_for(structure, sendBack, "sendBack");
+  structure->migrate(new_element, new_process, dist);
+
+  failures = kkLidView("fails", 1);
+  pids = structure->get<0>();
+  rnks = structure->get<3>();
+  auto checkPostBackMigrate = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
+    if (mask) {
+      if (rnks(p) != local_rank) {
+        printf("[ERROR] Test %s: Particle %d from rank %d was not sent back on rank %d\n",
+               name, pids(p), rnks(p), local_rank);
+        failures(0) = 1;
+      }
+    }
+  };
+  ps::parallel_for(structure, checkPostBackMigrate, "checkPostBackMigrate");
+
+  fails += ps::getLastValue<lid_t>(failures);
+  if (num_ptcls != structure->nPtcls()) {
+    printf("[ERROR] Test %s: Structure does not have all of the particles it started with on rank %d\n", name, comm_rank);
+    ++fails;
+  }
+
+  return fails;
+}
+
+int migrateToEmptyAndRefill(const char* name, PS* structure) {
+  printf("migrateToEmptyAndRefill %s, rank %d\n", name, comm_rank);
+
+  int fails = 0;
+  kkLidView failures("fails", 1);
+
+  int originalPtcls = structure->nPtcls();
+
+  kkLidView new_element("new_element", structure->capacity());
+  kkLidView new_process("new_process", structure->capacity());
+
+  int local_rank = comm_rank;
+  int local_csize = comm_size;
+  auto rnks = structure->get<3>();
+  auto elem = structure->get<2>();
+  int num_elems = structure->nElems();
+
+  auto sendToOdd = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
+    rnks(p) = local_rank;
+    elem(p) = e;
+    if (mask) {
+      if (local_rank % 2 == 0) {
+        new_process(p) = (local_rank + 1) % local_csize;
+        new_element(p) = num_elems - 1;
+      }
+      else {
+        new_process(p) = local_rank;
+        new_element(p) = e;
+      }
+    }
+    else {
+      new_element(p) = e;
+      new_process(p) = local_rank;
+    }
+  };
+  ps::parallel_for(structure, sendToOdd, "sendToOdd");
+  structure->migrate(new_element, new_process);
+
+  if (comm_rank % 2 == 0 && (comm_rank != 0 || comm_size % 2 == 0)) {
+    if (structure->nPtcls() != 0) {
+      ++fails;
+      fprintf(stderr, "[ERROR] Particles remain on rank %d\n", comm_rank);
+    }
+    auto checkPtcls = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
+      if (mask) {
+        failures(0) = 1;
+        printf("[ERROR] Particle %d remains on rank %d\n", p, local_rank);
+      }
+    };
+    ps::parallel_for(structure, checkPtcls, "checkPtcls");
+  }
+  else {
+    if (structure->nPtcls() < originalPtcls) {
+      ++fails;
+      fprintf(stderr, "[ERROR] No particles on rank %d\n", comm_rank);
+    }
+    const int prev_rank = ((comm_rank - 1) + comm_size) % comm_size;
+    auto new_rnks = structure->get<3>();
+    auto checkPtcls = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
+      if (mask) {
+        if (new_rnks(p) != local_rank && new_rnks(p) !=  prev_rank) {
+          failures(0) = 1;
+          printf("[ERROR] Particle %d is not from ranks %d or %d\n", p, local_rank, prev_rank);
+        }
+      }
+    };
+    ps::parallel_for(structure, checkPtcls, "checkPtcls");
+  }
+
+  //Send Particles back to original process
+  rnks = structure->get<3>();
+  new_element = kkLidView("new_element", structure->capacity());
+  new_process = kkLidView("new_process", structure->capacity());
+  auto sendToOrig = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
+    new_element(p) = e;
+    if (mask) {
+      new_process(p) = rnks(p);
+    }
+    else {
+      new_process(p) = local_rank;
+    }
+  };
+  ps::parallel_for(structure, sendToOrig, "sendToOrig");
+  structure->migrate(new_element, new_process);
+
+  if (structure->nPtcls() != originalPtcls) {
+    ++fails;
+    fprintf(stderr, "[ERROR] Number of particles does not match original on "
+            "rank %d [%d != %d]\n", comm_rank, structure->nPtcls(), originalPtcls);
+  }
+
+  auto elems = structure->get<2>();
+  new_element = kkLidView("new_element", structure->capacity());
+  auto resetElements = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
+    if (mask) {
+      new_element(p) = elems(p);
+    }
+    else {
+      new_element(p) = e;
+    }
+  };
+  structure->rebuild(new_element);
+
+  fails += pumipic::getLastValue<lid_t>(failures);
+  return fails;
+}
