+++ particle_structs/test/test_structure.cpp
@@ -1,7 +1,8 @@
 #include <particle_structs.hpp>
 #include "read_particles.hpp"
 #include <cmath>
-#include "test_rebuild.cpp"
+
+int comm_rank, comm_size;
 
 #ifdef PP_USE_CUDA
 typedef Kokkos::CudaSpace DeviceSpace;
@@ -13,7 +14,6 @@ void finalize() {
   MPI_Finalize();
 }
 
-int comm_rank, comm_size;
 //Structure adding functions
 int addSCSs(std::vector<PS*>& structures, std::vector<std::string>& names,
             lid_t num_elems, lid_t num_ptcls, kkLidView ppe,
@@ -25,7 +25,7 @@ int addCabMs(std::vector<PS*>& structures, std::vector<std::string>& names,
             lid_t num_elems, lid_t num_ptcls, kkLidView ppe,
             kkGidView element_gids, kkLidView particle_elements, PS::MTVs particle_info);
 
-//Simple tests of construction
+//Simple tests of constructors
 int testCounts(const char* name, PS* structure, lid_t num_elems, lid_t num_ptcls);
 int testParticleExistence(const char* name, PS* structure, lid_t num_ptcls);
 int setValues(const char* name, PS* structure);
@@ -91,9 +91,6 @@ int main(int argc, char* argv[]) {
     fails += addCabMs(structures, names, num_elems, num_ptcls, ppe, element_gids,
                      particle_elements, particle_info);
 
-
-
-
     //Run each structure on every test
     for (int i = 0; i < structures.size(); ++i) {
       fails += testCounts(names[i].c_str(), structures[i], num_elems, num_ptcls);
@@ -129,6 +126,7 @@ int main(int argc, char* argv[]) {
   return total_fails;
 
 }
+
 
 int addSCSs(std::vector<PS*>& structures, std::vector<std::string>& names,
             lid_t num_elems, lid_t num_ptcls, kkLidView ppe,
@@ -207,140 +205,6 @@ int addCabMs(std::vector<PS*>& structures, std::vector<std::string>& names,
 }
 
 
-int testCounts(const char* name, PS* structure, lid_t num_elems, lid_t num_ptcls) {
-  printf("testCounts %s\n", name);
-
-  int fails = 0;
-  if (structure->nElems() != num_elems) {
-    fprintf(stderr, "[ERROR] Test %s: Element count mismatch on rank %d "
-            "[(structure)%d != %d(actual)]\n", name,
-            comm_rank, structure->nElems(), num_elems);
-    ++fails;
-  }
-  if (structure->nPtcls() != num_ptcls) {
-    fprintf(stderr, "[ERROR] Test %s: Particle count mismatch on rank %d "
-            "[(structure)%d != %d(actual)]\n", name,
-            comm_rank, structure->nPtcls(), num_ptcls);
-    ++fails;
-  }
-  if (structure->numRows() < num_elems) {
-    fprintf(stderr, "[ERROR] Test %s: Number of rows is too small to fit elements on rank %d "
-            "[(structure)%d < %d(actual)]\n", name, comm_rank,
-            structure->numRows(), num_elems);
-    ++fails;
-  }
-  if (structure->capacity() < num_ptcls) {
-    fprintf(stderr, "[ERROR] Test %s: Capcity is too small to fit particles on rank %d "
-            "[(structure)%d < %d(actual)]\n", name, comm_rank,
-            structure->capacity(), num_ptcls);
-    ++fails;
-  }
-  return fails;
-}
-
-int testParticleExistence(const char* name, PS* structure, lid_t num_ptcls) {
-  printf("testParticleExistence %s, rank %d\n", name, comm_rank);
-
-  int fails = 0;
-  kkLidView count("count", 1);
-  auto checkExistence = PS_LAMBDA(const lid_t& e, const lid_t& p, const bool& mask) {
-    Kokkos::atomic_fetch_add(&(count(0)), mask);
-  };
-  ps::parallel_for(structure, checkExistence, "check particle existence");
-  lid_t c = ps::getLastValue<lid_t>(count);
-  if (c != num_ptcls) {
-    fprintf(stderr, "[ERROR] Test %s: Number of particles found in parallel_for "
-            "does not match the number of particles on rank %d"
-            "[(parallel_for)%d != %d(actual)]]n", name, comm_rank,
-            c, num_ptcls);
-    ++fails;
-  }
-  return fails;
-}
-
-int setValues(const char* name, PS* structure) {
-  printf("setValues %s, rank %d\n", name, comm_rank);
-
-  int fails = 0;
-  auto dbls = structure->get<1>();
-  auto bools = structure->get<2>();
-  auto nums = structure->get<3>();
-  int local_rank = comm_rank;
-  auto setValues = PS_LAMBDA(const lid_t& e, const lid_t& p, const bool& mask) {
-    if (mask) {
-      dbls(p, 0) = p * e * 100.0;
-      dbls(p, 1) = M_PI * p + M_PI / 2.0;
-      dbls(p, 2) = M_E * 2.5;
-      nums(p) = local_rank;
-      bools(p) = true;
-    }
-    else {
-      dbls(p, 0) = 0;
-      dbls(p, 1) = 0;
-      dbls(p, 2) = 0;
-      nums(p) = -1;
-      bools(p) = false;
-    }
-  };
-  Kokkos::fence();
-  Kokkos::Timer timer;
-  ps::parallel_for(structure, setValues, "setValues");
-  Kokkos::fence();
-  double time = timer.seconds();
-  printf("Time to set values %s : %f\n", name, time);
-  return fails;
-}
-
-int pseudoPush(const char* name, PS* structure) {
-  printf("pseudoPush %s, rank %d\n", name, comm_rank);
-
-  int fails = 0;
-
-  int elements = structure->nElems();
-  fprintf(stderr, "elements : %d\n", elements);
-  Kokkos::View<double*> parentElmData("parentElmData", elements);
-  fprintf(stderr, "parent elm data size : %d\n", parentElmData.size());
-  Kokkos::parallel_for("parentElmData", parentElmData.size(),
-      KOKKOS_LAMBDA(const lid_t& e){
-    parentElmData(e) = 2+3*e;
-  });
-  //pumipic::printView(parentElmData);
-
-  auto dbls = structure->get<1>();
-  auto bools = structure->get<2>();
-  auto nums = structure->get<3>();
-  int local_rank = comm_rank;
-  auto quickMaths = PS_LAMBDA(const lid_t& e, const lid_t& p, const bool& mask) {
-    //printf("e: %d\tp: %d\tmask: %d\n", e, p, mask);
-    if (mask) {
-      dbls(p, 0) += 10;
-      dbls(p, 1) += 10;
-      dbls(p, 2) += 10;
-      dbls(p, 0) = dbls(p,0) * dbls(p,0) * dbls(p,0) / sqrt(p) / sqrt(e) + parentElmData(e);
-      dbls(p, 1) = dbls(p,1) * dbls(p,1) * dbls(p,1) / sqrt(p) / sqrt(e) + parentElmData(e);
-      dbls(p, 2) = dbls(p,2) * dbls(p,2) * dbls(p,2) / sqrt(p) / sqrt(e) + parentElmData(e);
-      nums(p) = local_rank;
-      bools(p) = true;
-    }
-    else {
-      dbls(p, 0) = 0;
-      dbls(p, 1) = 0;
-      dbls(p, 2) = 0;
-      nums(p) = -1;
-      bools(p) = false;
-    }
-  };
-
-  Kokkos::fence();
-  Kokkos::Timer timer;
-  ps::parallel_for(structure, quickMaths, "setValues");
-  Kokkos::fence();
-  double time = timer.seconds();
-  printf("Time for math Ops on %s : %f\n", name, time);
-
-  return fails;
-}
-
 //Functionality tests
 int testRebuild(const char* name, PS* structure) {
   printf("testRebuild %s, rank %d\n", name, comm_rank);
@@ -352,100 +216,6 @@ int testRebuild(const char* name, PS* structure) {
   fails += rebuildNewPtcls(name, structure);
   fails += rebuildPtclsDestroyed(name, structure);
   fails += rebuildNewAndDestroyed(name, structure);
-
-  return fails;
-}
-
-int testMigration(const char* name, PS* structure) {
-  printf("testMigration %s, rank %d\n", name, comm_rank);
-
-  int fails = 0;
-  kkLidView failures("fails", 1);
-
-  kkLidView new_element("new_element", structure->capacity());
-  kkLidView new_process("new_process", structure->capacity());
-
-  int num_ptcls = structure->nPtcls();
-  //Send even particles one process to the right
-  auto pids = structure->get<0>();
-  auto rnks = structure->get<3>();
-  int local_rank = comm_rank;
-  int local_csize = comm_size;
-  int num_elems = structure->nElems();
-  auto sendRight = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
-    if (mask) {
-      new_element(p) = e;
-      if (e == num_elems - 1) {
-        new_process(p) = (local_rank + 1) % local_csize;
-      }
-      else
-        new_process(p) = local_rank;
-      rnks(p) = local_rank;
-    }
-    else {
-      new_element(p) = e;
-      new_process(p) = local_rank;
-    }
-  };
-  ps::parallel_for(structure, sendRight, "sendRight");
-  structure->migrate(new_element, new_process);
-
-  pids = structure->get<0>();
-  rnks = structure->get<3>();
-  auto checkPostMigrate = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
-    if (mask) {
-      const int pid = pids(p);
-      const int rank = rnks(p);
-      if (e == num_elems - 1 && rank == local_rank) {
-        printf("[ERROR] Failed to send particle %d on rank %d\n",
-               pid, local_rank);
-        failures(0) = 1;
-      }
-      if (e != 0 && rank != local_rank) {
-        printf("[ERROR] Incorrectly received particle %d from rank %d to rank %d in element %d\n", pid, rank, local_rank, e);
-        failures(0) = 1;
-      }
-    }
-  };
-  if (comm_size > 1)
-    ps::parallel_for(structure, checkPostMigrate, "checkPostMigrate");
-  fails += ps::getLastValue<lid_t>(failures);
-
-  //Make a distributor
-  int neighbors[3];
-  neighbors[0] = comm_rank;
-  neighbors[1] = (comm_rank - 1 + comm_size) % comm_size;
-  neighbors[2] = (comm_rank + 1) % comm_size;
-  ps::Distributor<typename PS::memory_space> dist(std::min(comm_size, 3), neighbors);
-
-  new_element = kkLidView("new_element", structure->capacity());
-  new_process = kkLidView("new_process", structure->capacity());
-  auto sendBack = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
-    new_element(p) = e;
-    new_process(p) = rnks(p);
-  };
-  ps::parallel_for(structure, sendBack, "sendBack");
-  structure->migrate(new_element, new_process, dist);
-
-  failures = kkLidView("fails", 1);
-  pids = structure->get<0>();
-  rnks = structure->get<3>();
-  auto checkPostBackMigrate = PS_LAMBDA(const lid_t e, const lid_t p, const bool mask) {
-    if (mask) {
-      if (rnks(p) != local_rank) {
-        printf("[ERROR] Test %s: Particle %d from rank %d was not sent back on rank %d\n",
-               name, pids(p), rnks(p), local_rank);
-        failures(0) = 1;
-      }
-    }
-  };
-  ps::parallel_for(structure, checkPostBackMigrate, "checkPostBackMigrate");
-
-  fails += ps::getLastValue<lid_t>(failures);
-  if (num_ptcls != structure->nPtcls()) {
-    printf("[ERROR] Test %s: Structure does not have all of the particles it started with on rank %d\n", name, comm_rank);
-    ++fails;
-  }
 
   return fails;
 }
@@ -588,109 +358,6 @@ int testSegmentComp(const char* name, PS* structure) {
   return fails;
 }
 
-int migrateToEmptyAndRefill(const char* name, PS* structure) {
-  printf("migrateToEmptyAndRefill %s, rank %d\n", name, comm_rank);
-
-  int fails = 0;
-  kkLidView failures("fails", 1);
-
-  int originalPtcls = structure->nPtcls();
-
-  kkLidView new_element("new_element", structure->capacity());
-  kkLidView new_process("new_process", structure->capacity());
-
-  int local_rank = comm_rank;
-  int local_csize = comm_size;
-  auto rnks = structure->get<3>();
-  auto elem = structure->get<2>();
-  int num_elems = structure->nElems();
-
-  auto sendToOdd = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
-    rnks(p) = local_rank;
-    elem(p) = e;
-    if (mask) {
-      if (local_rank % 2 == 0) {
-        new_process(p) = (local_rank + 1) % local_csize;
-        new_element(p) = num_elems - 1;
-      }
-      else {
-        new_process(p) = local_rank;
-        new_element(p) = e;
-      }
-    }
-    else {
-      new_element(p) = e;
-      new_process(p) = local_rank;
-    }
-  };
-  ps::parallel_for(structure, sendToOdd, "sendToOdd");
-  structure->migrate(new_element, new_process);
-
-  if (comm_rank % 2 == 0 && (comm_rank != 0 || comm_size % 2 == 0)) {
-    if (structure->nPtcls() != 0) {
-      ++fails;
-      fprintf(stderr, "[ERROR] Particles remain on rank %d\n", comm_rank);
-    }
-    auto checkPtcls = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
-      if (mask) {
-        failures(0) = 1;
-        printf("[ERROR] Particle %d remains on rank %d\n", p, local_rank);
-      }
-    };
-    ps::parallel_for(structure, checkPtcls, "checkPtcls");
-  }
-  else {
-    if (structure->nPtcls() < originalPtcls) {
-      ++fails;
-      fprintf(stderr, "[ERROR] No particles on rank %d\n", comm_rank);
-    }
-    const int prev_rank = ((comm_rank - 1) + comm_size) % comm_size;
-    auto new_rnks = structure->get<3>();
-    auto checkPtcls = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
-      if (mask) {
-        if (new_rnks(p) != local_rank && new_rnks(p) !=  prev_rank) {
-          failures(0) = 1;
-          printf("[ERROR] Particle %d is not from ranks %d or %d\n", p, local_rank, prev_rank);
-        }
-      }
-    };
-    ps::parallel_for(structure, checkPtcls, "checkPtcls");
-  }
-
-  //Send Particles back to original process
-  rnks = structure->get<3>();
-  new_element = kkLidView("new_element", structure->capacity());
-  new_process = kkLidView("new_process", structure->capacity());
-  auto sendToOrig = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
-    new_element(p) = e;
-    if (mask) {
-      new_process(p) = rnks(p);
-    }
-    else {
-      new_process(p) = local_rank;
-    }
-  };
-  ps::parallel_for(structure, sendToOrig, "sendToOrig");
-  structure->migrate(new_element, new_process);
-
-  if (structure->nPtcls() != originalPtcls) {
-    ++fails;
-    fprintf(stderr, "[ERROR] Number of particles does not match original on "
-            "rank %d [%d != %d]\n", comm_rank, structure->nPtcls(), originalPtcls);
-  }
-
-  auto elems = structure->get<2>();
-  new_element = kkLidView("new_element", structure->capacity());
-  auto resetElements = PS_LAMBDA(lid_t e, lid_t p, bool mask) {
-    if (mask) {
-      new_element(p) = elems(p);
-    }
-    else {
-      new_element(p) = e;
-    }
-  };
-  structure->rebuild(new_element);
-
-  fails += pumipic::getLastValue<lid_t>(failures);
-  return fails;
-}
+#include "test_constructor.cpp"
+#include "test_rebuild.cpp"
+#include "test_migrate.cpp"
